Cleaning up some compiler warnings

+++ src/EngineCore/CameraComponent.cpp
@@ -49,14 +49,14 @@ namespace EngineCore
             return m_active_camera;
         }
 
-        Entity CameraComponentManager::getEntity(uint index) const
+        Entity CameraComponentManager::getEntity(size_t index) const
         {
             auto indices = data_.getIndices(index);
 
             return data_(indices.first, indices.second).entity;
         }
 
-        void CameraComponentManager::setCameraAttributes(uint index, float near_cp, float far_cp, float fovy, float aspect_ratio, float exposure)
+        void CameraComponentManager::setCameraAttributes(size_t index, float near_cp, float far_cp, float fovy, float aspect_ratio, float exposure)
         {
             auto indices = data_.getIndices(index);
             
@@ -69,7 +69,7 @@ namespace EngineCore
             updateProjectionMatrix(index);
         }
 
-        void CameraComponentManager::updateProjectionMatrix(uint index)
+        void CameraComponentManager::updateProjectionMatrix(size_t index)
         {
             auto indices = data_.getIndices(index);
 
@@ -102,55 +102,55 @@ namespace EngineCore
             //    m_data.projection_matrix[index] = projection_matrix;
         }
 
-        void CameraComponentManager::setNear(uint index, float near_cp)
+        void CameraComponentManager::setNear(size_t index, float near_cp)
         {
             auto indices = data_.getIndices(index);
             data_(indices.first, indices.second).near_cp = near_cp;
         }
 
-        void CameraComponentManager::setFar(uint index, float far_cp)
+        void CameraComponentManager::setFar(size_t index, float far_cp)
         {
             auto indices = data_.getIndices(index);
             data_(indices.first, indices.second).far_cp = far_cp;
         }
 
-        Mat4x4 CameraComponentManager::getProjectionMatrix(uint index) const
+        Mat4x4 CameraComponentManager::getProjectionMatrix(size_t index) const
         {
             auto indices = data_.getIndices(index);
             return data_(indices.first, indices.second).projection_matrix;
         }
 
-        float CameraComponentManager::getFovy(uint index) const
+        float CameraComponentManager::getFovy(size_t index) const
         {
             auto indices = data_.getIndices(index);
             return data_(indices.first, indices.second).fovy;
         }
 
-        void CameraComponentManager::setFovy(uint index, float fovy)
+        void CameraComponentManager::setFovy(size_t index, float fovy)
         {
             auto indices = data_.getIndices(index);
             data_(indices.first, indices.second).fovy = fovy;
         }
 
-        float CameraComponentManager::getAspectRatio(uint index) const
+        float CameraComponentManager::getAspectRatio(size_t index) const
         {
             auto indices = data_.getIndices(index);
             return data_(indices.first, indices.second).aspect_ratio;
         }
 
-        void CameraComponentManager::setAspectRatio(uint index, float aspect_ratio)
+        void CameraComponentManager::setAspectRatio(size_t index, float aspect_ratio)
         {
             auto indices = data_.getIndices(index);
             data_(indices.first, indices.second).aspect_ratio = aspect_ratio;
         }
 
-        float CameraComponentManager::getExposure(uint index) const
+        float CameraComponentManager::getExposure(size_t index) const
         {
             auto indices = data_.getIndices(index);
             return data_(indices.first, indices.second).exposure;
         }
 
-        void CameraComponentManager::setExposure(uint index, float exposure)
+        void CameraComponentManager::setExposure(size_t index, float exposure)
         {
             auto indices = data_.getIndices(index);
             data_(indices.first, indices.second).exposure = exposure;
